Extract PageHeader.Title, PageHeader.Image, PageHeader.CTA into standalone components
