+++ apps/ffdweb-site/src/app/learning-resources/page.tsx
@@ -11,12 +11,14 @@ import { graphicsData } from '@/data/graphicsData'
 import { createMetadata } from '@/utils/createMetadata'
 
 import { CTASection } from '@/components/CTASection'
-import { PageHeader } from '@/components/PageHeader'
+
 
 import { LearningResourcesContent } from './components/LearningResourcesContent'
 import { LEARNING_RESOURCES_SEO } from './constants/seo'
 import { generateStructuredData } from './utils/generateStructuredData'
 import { getLearningResourcesData } from './utils/getLearningResourceData'
+
+import { PageHeader } from '@/_components/PageHeader/PageHeader'
 
 export default async function LearningResources() {
   const resources = await getLearningResourcesData()
