+++ apps/ffdweb-site/src/app/projects/page.tsx
@@ -10,12 +10,14 @@ import { graphicsData } from '@/data/graphicsData'
 import { createMetadata } from '@/utils/createMetadata'
 
 import { ContactCTA } from '@/components/ContactCTA'
-import { PageHeader } from '@/components/PageHeader'
+
 
 import { ProjectsContent } from './components/ProjectsContent'
 import { PROJECTS_SEO } from './constants/seo'
 import { generateStructuredData } from './utils/generateStructuredData'
 import { getProjectsData } from './utils/getProjectData'
+
+import { PageHeader } from '@/_components/PageHeader/PageHeader'
 
 export default async function Projects() {
   const projects = await getProjectsData()
@@ -25,7 +27,7 @@ export default async function Projects() {
       <StructuredDataScript
         structuredData={generateStructuredData(PROJECTS_SEO)}
       />
-      
+
       <PageHeader
         kicker="Projects"
         title="Discover Current and Past FFDW Project Partners"
